common makeSelfEdge: take 'ind' as a 'size_t'
Pays down some cast noise from prior commits.

+++ lib/common/render.h
@@ -113,7 +113,7 @@ extern "C" {
     RENDER_API void* init_xdot (Agraph_t* g);
     RENDER_API int initMapData (GVJ_t*, char*, char*, char*, char*, char*, void*);
     RENDER_API bool isPolygon(node_t *);
-    RENDER_API void makeSelfEdge(edge_t *edges[], int ind, size_t cnt,
+    RENDER_API void makeSelfEdge(edge_t *edges[], size_t ind, size_t cnt,
 	double sizex, double sizey, splineInfo * sinfo);
     RENDER_API textlabel_t *make_label(void *obj, char *str, int kind, double fontsize, char *fontname, char *fontcolor);
     RENDER_API bezier *new_spline(edge_t *e, size_t sz);
